+++ NEWS.md
@@ -3,6 +3,7 @@
 ## New features
 
 * Exporting to `cytoscape()` for visualization.
+* Added caching for `neo4j_engine` queries (enabled by default, length of R session only)
 
 ## Bug fixes